moving app parts to atomic structure, needs some rework yet
