--- src/components/layout/BackgroundAnimation/index.tsx
@@ -1,5 +1,0 @@
-import styles from './styles.module.scss'
-
-export default function BackgroundAnimation () {
-  return <div className={ styles.container } />
-}
