--- src/components/layout/BackgroundAnimation/styles.module.scss
@@ -1,60 +1,0 @@
-.container {
-  width: 100%;
-  height: 100%;
-  animation: background-animate 10s cubic-bezier(0.445, 0.05, 0.55, 0.95) infinite;
-  background-image: url('https://images.unsplash.com/photo-1663970206579-c157cba7edda?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=120&q=10');
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;
-  border-radius: 50%;
-  position: fixed;
-  top: 0;
-  left: 0;
-
-  &::after {
-    content: '';
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 30rem;
-    height: 20rem;
-    opacity: 0.5;
-    background: #9f2ac8;
-    border-radius: 20%;
-    animation: background-animate 8s cubic-bezier(0.445, 0.05, 0.55, 0.95) infinite,
-      around 3s cubic-bezier(0.445, 0.05, 0.55, 0.95) infinite reverse;
-  }
-
-  &::before {
-    content: '';
-    position: absolute;
-    top: 0;
-    left: 50%;
-    width: 10rem;
-    height: 70rem;
-    background: #9e6bff;
-    border-radius: 50%;
-    transform: translateX(10rem);
-    animation: background-animate 5s cubic-bezier(0.445, 0.05, 0.55, 0.95) infinite,
-      around 5s cubic-bezier(0.445, 0.05, 0.55, 0.95) infinite;
-  }
-}
-
-@keyframes background-animate {
-  0% {
-    transform: rotate3d(0, 0, 10, 0deg);
-  }
-  100% {
-    transform: rotate3d(0, 0, 10, 360deg);
-  }
-}
-
-@keyframes around {
-  to {
-    transform: translate3d(10, 20, 0) rotate3d(0, 0, 1, 360deg);
-  }
-
-  from {
-    transform: translate3d(10, 20, 30) rotate3d(0, 0, 10, 0deg);
-  }
-}
