+++ src/components/molecules/Sidebar/styles.module.scss
