+++ src/pages/app/index.tsx
@@ -1,22 +1,9 @@
 import { useSession } from 'next-auth/react'
 import { useRouter } from 'next/router'
-import type { NextPageWithLayout } from '../_app'
+import { FiPlus } from 'react-icons/fi'
+import type { NextPageWithLayout } from '@/pages/_app'
 import App from '@/layouts/App'
 import { TaskProvider } from '@/contexts/task'
-import { FiPlus } from 'react-icons/fi'
-
-const styles = {
-  container: {
-    width: '100%',
-    height: '100%',
-    minHeight: '100vh',
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-    fontSize: 'xxx-large'
-  } as React.CSSProperties,
-}
 
 const AppPage: NextPageWithLayout = () => {
   const { status } = useSession()
@@ -24,7 +11,7 @@ const AppPage: NextPageWithLayout = () => {
 
   if (status === 'loading') {
     return (
-      <div style={ styles.container }>
+      <div>
         <h1>Loading...</h1>
       </div>
     )
@@ -34,8 +21,7 @@ const AppPage: NextPageWithLayout = () => {
     push('/signin')
 
     return (
-      <div
-        style={ styles.container }>
+      <div>
         <p>You are not signed in.</p>
         <p>Redirecting to Sign In...</p>
       </div>
@@ -45,12 +31,10 @@ const AppPage: NextPageWithLayout = () => {
   return (
     <TaskProvider>
       <App>
-        <div style={ styles.container }>
+        <div>
           <h1>App Page</h1>
           <p>When user is Logged!</p>
-          <label
-            style={{ display: 'flex', backgroundColor: 'white', padding: '1rem', borderRadius: '0.5rem' }}
-          >
+          <label>
             <FiPlus />
             <input type='text' placeholder='Add a task' />
           </label>
